pyproject.toml
[project]
name = "purview-mcp"
version = "0.1.0"
description = "Microsoft Purview MCP Server for Claude Code integration"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0",
    "azure-identity>=1.16",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "databricks-sdk>=0.20",
]

[project.scripts]
purview-mcp = "purview_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/purview_mcp"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
